pyproject.toml
[project]
name = "SlurmSlim"
version = "0.1.0"
description = "Optimize Slurm Job Scheduling with MCP-Based Memory Estimation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "httpx[socks]>=0.28.1",
    "mcp[cli]>=1.4.1",
    "memory-profiler>=0.61.0",
    "mistune>=3.1.2",
    "openai>=1.66.3",
    "python-dotenv>=1.0.1",
]
